app/core/jwt.py
from datetime import datetime, timedelta
from jose import jwt
from app.core.config import TOKEN_EXPIRATION_AFTER, SECRET_KEY, ALGORITHM

def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=TOKEN_EXPIRATION_AFTER))
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt