requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt requirements.in
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.11.0
    # via
    #   -r requirements.in
    #   starlette
click==8.3.0
    # via
    #   -r requirements.in
    #   uvicorn
dnspython==2.8.0
    # via
    #   -r requirements.in
    #   email-validator
email-validator==2.3.0
    # via -r requirements.in
fastapi==0.120.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   -r requirements.in
    #   uvicorn
idna==3.11
    # via
    #   -r requirements.in
    #   anyio
    #   email-validator
mariadb==1.1.14
    # via -r requirements.in
packaging==25.0
    # via mariadb
pillow==12.0.0
    # via -r requirements.in
pydantic==2.12.3
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.4
    # via
    #   -r requirements.in
    #   pydantic
pydantic-settings==2.11.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
sqlalchemy==2.0.44
    # via -r requirements.in
starlette==0.48.0
    # via
    #   -r requirements.in
    #   fastapi
typing-extensions==4.15.0
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   -r requirements.in
    #   pydantic
    #   pydantic-settings
uv==0.9.5
    # via -r requirements.in
uvicorn==0.38.0
    # via -r requirements.in
python-jose[cryptography]==3.3.0
